decl.c (gfc_match_old_kind_spec,match_type_spec): Use gfc_std_notify to report nonstandard intrinsic type declarations.
2005-12-17  Steven G. Kargl  <kargls@comcast.net>

	* decl.c (gfc_match_old_kind_spec,match_type_spec): Use gfc_std_notify
	to report nonstandard intrinsic type declarations.

	* gfortran.dg/imag_1.f: Fix nonstandard type declarations.
	* gfortran.dg/nested_modules_1.f90: Ditto.
	* gfortran.dg/hollerith_f95.f90: Ditto.
	* gfortran.dg/select_5.f90: Ditto.
	* gfortran.dg/secnds.f: Ditto.
	* gfortran.dg/hollerith2.f90: Ditto.
	* gfortran.dg/imag_2.f: Ditto.
	* gfortran.dg/ftell_2.f90: Ditto.
	* gfortran.dg/malloc_free_1.f90: Ditto.
	* gfortran.dg/logint-1.f: Ditto.
	* gfortran.dg/recursive_statement_functions.f90: Ditto.
	* gfortran.dg/e_d_fmt.f90: Ditto.
	* gfortran.dg/hollerith_legacy.f90: Ditto.
	* gfortran.dg/logint-2.f: Ditto.
	* gfortran.dg/enum_5.f90: Ditto.
	* gfortran.dg/f2c_2.f90: Ditto.
	* gfortran.dg/pr17143.f90: Ditto.
	* gfortran.dg/namelist_14.f90: Ditto.
	* gfortran.dg/logint-3.f: Ditto.
	* gfortran.dg/spread_scalar_source.f90: Ditto.
	* gfortran.dg/fmt_read_bz_bn.f90: Ditto.
	* gfortran.dg/namelist_11.f: Ditto.
	* gfortran.dg/g77/intrinsic-unix-bessel.f: Ditto.
	* gfortran.dg/g77/20010519-1.f
	* gfortran.dg/g77/alpha1.f: Ditto.
	* gfortran.dg/g77/990115-1.f: Ditto.
	* gfortran.dg/g77/erfc.f: Ditto.
	* gfortran.dg/g77/19990313-3.f: Ditto.
	* gfortran.dg/g77/f90-intrinsic-numeric.f: Ditto.
	* gfortran.dg/g77/20010426.f: Ditto.
	* gfortran.dg/g77/19990313-0.f: Ditto.
	* gfortran.dg/g77/f90-intrinsic-mathematical.f: Ditto.
	* gfortran.dg/g77/20000629-1.f: Ditto.
	* gfortran.dg/g77/970125-0.f: Ditto.
	* gfortran.dg/g77/8485.f: Ditto.
	* gfortran.dg/g77/f90-intrinsic-bit.f: Ditto.
	* gfortran.dg/g77/19990313-1.f: Ditto.
	* gfortran.dg/g77/int8421.f: Ditto.
	* gfortran.dg/g77/19990305-0.f: Ditto.
	* gfortran.dg/g77/947.f: Ditto.
	* gfortran.dg/g77/19990905-2.f: Ditto.
	* gfortran.dg/g77/cabs.f: Ditto.
	* gfortran.dg/g77/19990313-2.f: Ditto.
	* gfortran.dg/g77/20020307-1.f: Ditto.
	* gfortran.dg/g77/dcomplex.f: Ditto.
	* gfortran.dg/g77/19990502-1.f: Ditto.
	* gfortran.dg/g77/19981216-0.f: Ditto.
	* gfortran.dg/common_4.f90: Ditto.
	* gfortran.dg/entry_4.f90: Ditto.
	* gfortran.dg/equiv_constraint_2.f90: Ditto.
	* gfortran.dg/g77_intrinsics_funcs.f: Ditto.
	* gfortran.dg/namelist_12.f: Ditto.
	* gfortran.dg/scale_1.f90: Ditto.
	* gfortran.dg/modulo_1.f90
	* gfortran.dg/hollerith.f90: Ditto.
	* gfortran.dg/direct_io_3.f90: Ditto.
	* gfortran.dg/ftell_1.f90: Ditto.: Ditto.
	* gfortran.dg/gnu_logical_1.F: Ditto.
	* gfortran.dg/unf_io_convert_1.f90: Ditto.
	* gfortran.dg/assign_1.f90: Ditto.
	* gfortran.dg/g77_intrinsics_sub.f: Ditto.
	* gfortran.dg/pr23095.f: Ditto.
	* gfortran.dg/read_float_1.f90: Ditto.

From-SVN: r108715

+++ gcc/testsuite/gfortran.dg/common_4.f90
@@ -4,8 +4,8 @@
 ! Check misaligned common blocks.
 program prog
   common /block/ a, b, c
-  integer*1 a
-  integer*4 b, c
+  integer(kind=1) a
+  integer b, c
   a = 1
   b = HUGE(b)
   c = 2
@@ -13,7 +13,7 @@ program prog
 end program
 subroutine foo
   common /block/ a, b, c
-  integer*1 a
-  integer*4 b, c
+  integer(kind=1) a
+  integer b, c
   if (a .ne. 1 .or. b .ne. HUGE(b) .or. c .ne. 2) call abort
 end subroutine
